+++ hive_generator/lib/src/adapter_builder/adapter_builder.dart
@@ -7,8 +7,6 @@ import 'package:meta/meta.dart';
 @immutable
 class AdapterField {
   /// The corresponding element for this field
-  /// TODO: Fix with analyzer 8
-  /// ignore: deprecated_member_use
   final PropertyAccessorElement element;
 
   /// The index of the field
@@ -43,8 +41,6 @@ class AdapterField {
 @immutable
 abstract class AdapterBuilder {
   /// TODO: Document this!
-  /// TODO: Fix with analyzer 8
-  /// ignore: deprecated_member_use
   final InterfaceElement cls;
 
   /// TODO: Document this!