+++ hive_generator/lib/src/adapter_builder/class_adapter_builder.dart
@@ -21,19 +21,24 @@ class ClassAdapterBuilder extends AdapterBuilder {
   );
 
   /// [TypeChecker] for [HiveList].
-  final hiveListChecker = const TypeChecker.fromRuntime(HiveList);
+  final hiveListChecker =
+      const TypeChecker.typeNamed(HiveList, inPackage: 'hive_ce');
 
   /// [TypeChecker] for [Map].
-  final mapChecker = const TypeChecker.fromRuntime(Map);
+  final mapChecker =
+      const TypeChecker.typeNamed(Map, inPackage: 'core', inSdk: true);
 
   /// [TypeChecker] for [Set].
-  final setChecker = const TypeChecker.fromRuntime(Set);
+  final setChecker =
+      const TypeChecker.typeNamed(Set, inPackage: 'core', inSdk: true);
 
   /// [TypeChecker] for [Iterable].
-  final iterableChecker = const TypeChecker.fromRuntime(Iterable);
+  final iterableChecker =
+      const TypeChecker.typeNamed(Iterable, inPackage: 'core', inSdk: true);
 
   /// [TypeChecker] for [Uint8List].
-  final uint8ListChecker = const TypeChecker.fromRuntime(Uint8List);
+  final uint8ListChecker = const TypeChecker.typeNamed(Uint8List,
+      inPackage: 'typed_data', inSdk: true);
 
   @override
   String buildRead() {
@@ -43,10 +48,10 @@ class ClassAdapterBuilder extends AdapterBuilder {
     final fields = setters.toList();
 
     // Empty classes
-    if (constr.parameters.isEmpty && fields.isEmpty) {
+    if (constr.formalParameters.isEmpty && fields.isEmpty) {
       return '''
     reader.readByte();
-    return ${cls.name}();
+    return ${cls.displayName}();
     ''';
     }
 
@@ -57,16 +62,16 @@ class ClassAdapterBuilder extends AdapterBuilder {
       for (int i = 0; i < numOfFields; i++)
         reader.readByte(): reader.read(),
     };
-    return ${cls.name}(
+    return ${cls.displayName}(
     ''');
 
-    for (final param in constr.parameters) {
-      var field = fields.firstWhereOrNull((it) => it.name == param.name);
+    for (final param in constr.formalParameters) {
+      var field = fields.firstWhereOrNull((it) => it.name == param.displayName);
       // Final fields
-      field ??= getters.firstWhereOrNull((it) => it.name == param.name);
+      field ??= getters.firstWhereOrNull((it) => it.name == param.displayName);
       if (field != null) {
         if (param.isNamed) {
-          code.write('${param.name}: ');
+          code.write('${param.displayName}: ');
         }
         code.write(_value(param.type, field));
         code.writeln(',');
@@ -125,8 +130,6 @@ class ClassAdapterBuilder extends AdapterBuilder {
     } else if (type.isDartCoreDouble) {
       return '($variable as num$suffix)$suffix.toDouble()';
     } else {
-      /// TODO: Fix with analyzer 8
-      /// ignore: deprecated_member_use
       return '$variable as ${type.getPrefixedDisplayString(cls.library)}';
     }
   }
@@ -158,8 +161,6 @@ class ClassAdapterBuilder extends AdapterBuilder {
 
       return '$suffix.map((e) => ${_cast(arg, 'e')})$cast';
     } else {
-      /// TODO: Fix with analyzer 8
-      /// ignore: deprecated_member_use
       return '$suffix.cast<${arg.getPrefixedDisplayString(cls.library)}>()';
     }
   }
@@ -173,12 +174,7 @@ class ClassAdapterBuilder extends AdapterBuilder {
       return '$suffix.map((dynamic k, dynamic v)=>'
           'MapEntry(${_cast(arg1, 'k')},${_cast(arg2, 'v')}))';
     } else {
-      /// TODO: Fix with analyzer 8
-      /// ignore: deprecated_member_use
       return '$suffix.cast<${arg1.getPrefixedDisplayString(cls.library)}, '
-
-          /// TODO: Fix with analyzer 8
-          /// ignore: deprecated_member_use
           '${arg2.getPrefixedDisplayString(cls.library)}>()';
     }
   }
@@ -223,31 +219,23 @@ String _suffixFromType(DartType type) {
 }
 
 extension on DartType {
-  /// TODO: Fix with analyzer 8
-  /// ignore: deprecated_member_use
   String getPrefixedDisplayString(LibraryElement currentLibrary) {
-    /// TODO: Fix with analyzer 8
-    /// ignore: deprecated_member_use
     final element = this.element;
     if (element == null) return getDisplayString();
 
-    /// TODO: Fix with analyzer 8
-    /// ignore: deprecated_member_use
     final definingLibrary = element.library;
     if (definingLibrary == currentLibrary) return getDisplayString();
 
-    /// TODO: Fix with analyzer 8
-    /// ignore: deprecated_member_use
-    for (final import in currentLibrary.units.expand((e) => e.libraryImports)) {
-      for (final MapEntry(:key, :value)
+    final prefix = currentLibrary.fragments
+        .expand((e) => e.libraryImports)
+        .firstWhereOrNull(
+            (e) => e.namespace.definedNames2.values.contains(element))
+        ?.prefix
+        ?.element
+        .displayName;
 
-          /// TODO: Fix with analyzer 8
-          /// ignore: deprecated_member_use
-          in import.namespace.definedNames.entries) {
-        if (value == element) {
-          return '$key${_suffixFromType(this)}';
-        }
-      }
+    if (prefix != null) {
+      return '$prefix.${getDisplayString()}';
     }
 
     return getDisplayString();