+++ hive_generator/lib/src/builder/schema_migrator_builder.dart
@@ -60,7 +60,7 @@ class SchemaMigratorBuilder implements Builder {
       if (!await buildStep.resolver.isLibrary(input)) continue;
       final library = await buildStep.resolver.libraryFor(input);
       final hiveTypeElements = LibraryReader(library)
-          .annotatedWith(TypeChecker.fromRuntime(HiveType));
+          .annotatedWith(TypeChecker.typeNamed(HiveType, inPackage: 'hive_ce'));
       hiveTypes.addAll(hiveTypeElements);
     }
 
@@ -69,8 +69,6 @@ class SchemaMigratorBuilder implements Builder {
       final cls = getClass(type.element);
       final className = cls.displayName;
 
-      /// TODO: Fix with analyzer 8
-      /// ignore: deprecated_member_use
       final library = type.element.library!;
       final typeId = readTypeId(type.annotation);
       final result = TypeAdapterGenerator.getAccessors(
@@ -93,15 +91,14 @@ class SchemaMigratorBuilder implements Builder {
         }
       }
 
-      final uri = library.source.uri;
+      final uri = library.uri;
       final isEnum = cls.thisType.isEnum;
       final constructor = getConstructor(cls);
       final accessors = [
-        ...cls.accessors,
-
-        /// TODO: Fix with analyzer 8
-        /// ignore: deprecated_member_use
-        ...cls.allSupertypes.expand((it) => it.accessors),
+        ...cls.getters,
+        ...cls.setters,
+        ...cls.allSupertypes
+            .expand((it) => [...it.element.getters, ...it.element.setters]),
       ];
       final info = _SchemaInfo(
         uri: uri,
@@ -192,13 +189,7 @@ class _SchemaInfo {
     required this.uri,
     required this.className,
     required bool isEnum,
-
-    /// TODO: Fix with analyzer 8
-    /// ignore: deprecated_member_use
     required ConstructorElement constructor,
-
-    /// TODO: Fix with analyzer 8
-    /// ignore: deprecated_member_use
     required List<PropertyAccessorElement> accessors,
     required HiveSchemaType schema,
   }) : schema = _sanitizeSchema(
@@ -213,13 +204,7 @@ class _SchemaInfo {
     required String className,
     required bool isEnum,
     required HiveSchemaType schema,
-
-    /// TODO: Fix with analyzer 8
-    /// ignore: deprecated_member_use
     required ConstructorElement constructor,
-
-    /// TODO: Fix with analyzer 8
-    /// ignore: deprecated_member_use
     required List<PropertyAccessorElement> accessors,
   }) {
     // Enums need no sanitization
@@ -231,12 +216,12 @@ class _SchemaInfo {
       final publicFieldName =
           fieldName.startsWith('_') ? fieldName.substring(1) : fieldName;
 
-      final isInConstructor =
-          constructor.parameters.any((e) => e.displayName == publicFieldName);
+      final isInConstructor = constructor.formalParameters
+          .any((e) => e.displayName == publicFieldName);
       final publicAccessors =
           accessors.where((e) => e.displayName == publicFieldName).toList();
-      final hasPublicSetter = publicAccessors.any((e) => e.isSetter);
-      final hasPublicGetter = publicAccessors.any((e) => e.isGetter);
+      final hasPublicSetter = publicAccessors.any((e) => e is SetterElement);
+      final hasPublicGetter = publicAccessors.any((e) => e is GetterElement);
 
       if (!isInConstructor && !hasPublicSetter) {
         throw InvalidGenerationSourceError(