+++ hive_generator/lib/src/helper/helper.dart
@@ -11,7 +11,8 @@ import 'package:path/path.dart' as path;
 import 'package:meta/meta.dart';
 import 'package:yaml_writer/yaml_writer.dart';
 
-final _hiveFieldChecker = const TypeChecker.fromRuntime(HiveField);
+final _hiveFieldChecker =
+    const TypeChecker.typeNamed(HiveField, inPackage: 'hive_ce');
 final _freezedDefaultChecker = const TypeChecker.fromUrl(
   'package:freezed_annotation/freezed_annotation.dart#Default',
 );
@@ -30,8 +31,6 @@ class HiveFieldInfo {
 }
 
 /// TODO: Document this!
-/// TODO: Fix with analyzer 8
-/// ignore: deprecated_member_use
 HiveFieldInfo? getHiveFieldAnn(Element? element) {
   if (element == null) return null;
   final obj = _hiveFieldChecker.firstAnnotationOfExact(element);
@@ -44,8 +43,6 @@ HiveFieldInfo? getHiveFieldAnn(Element? element) {
 }
 
 /// Get the string representation of the freezed default value
-/// TODO: Fix with analyzer 8
-/// ignore: deprecated_member_use
 DartObject? getFreezedDefault(Element? element) {
   if (element == null) return null;
   final obj = _freezedDefaultChecker.firstAnnotationOfExact(element);
@@ -56,30 +53,20 @@ DartObject? getFreezedDefault(Element? element) {
 }
 
 /// Get a classes default constructor or throw
-/// TODO: Fix with analyzer 8
-/// ignore: deprecated_member_use
 ConstructorElement getConstructor(InterfaceElement cls) {
-  /// TODO: Fix with analyzer 8
-  /// ignore: deprecated_member_use
-  final constr = cls.constructors.firstWhereOrNull((it) => it.name.isEmpty);
+  final constr = cls.constructors.firstWhereOrNull((it) => it.name == 'new');
   if (constr == null) {
     throw 'Provide an unnamed constructor.';
   }
   return constr;
 }
 
-/// TODO: Fix with analyzer 8
-/// ignore: deprecated_member_use
 /// Returns [element] as [InterfaceElement] if it is a class or enum
-/// TODO: Fix with analyzer 8
-/// ignore: deprecated_member_use
 InterfaceElement getClass(Element element) {
   if (element.kind != ElementKind.CLASS && element.kind != ElementKind.ENUM) {
     throw 'Only classes or enums are allowed to be annotated with @HiveType.';
   }
 
-  /// TODO: Fix with analyzer 8
-  /// ignore: deprecated_member_use
   return element as InterfaceElement;
 }